update host op test
